separate all logic and canvas elements for easier testing
