+++ server/server.js
@@ -17,9 +17,9 @@ const app = express();
 
 app.set("trust proxy", 1);
 
-app.use(cors({ origin: "https://ivysgame.netlify.app", credentials: true }));
+// app.use(cors({ origin: "https://ivysgame.netlify.app", credentials: true }));
 
-// app.use(cors());
+app.use(cors());
 app.options("*", cors());
 app.use(express.json({ limit: "100mb" }));
 app.use(
